Add 'Type-check name' from the book
The book 'Type-driven Development with Idris' by Edwin Brady
describes a 'Type-check name' command in the Atom extension
for Idris (called with Ctrl-Alt-T).

It is described as showing the type of a hole, and the types
of all the variables associated with it.

It seemed useful to me but there wasn't tooling for it, so
I have attempted to replicate its functionality as close
to the description in the book as I could get it.

Assuming the cursor is over a hole,
`idris2.metavars.type_check()` creates a popup containing
1. The names and types of the premises
2. A seperator line
3. The name and type of the hole the cursor is over